requirements.txt
django~=3.0.4
djangorestframework~=3.12.1
django-filter~=2.4.0
django-debug-toolbar~=2.2
pymysql~=0.9.2
simpleui